requirements.txt
customtkinter==5.2.2
tkinterdnd2==0.3.0
pillow==10.2.0
